+++ netharn/data/transforms/augmenters.py
@@ -194,14 +194,21 @@ class Resize(augmenter_base.ParamatarizedAugmenter):
         target_size (tuple): Scale images to this size (w, h) keeping the
         aspect ratio using a letterbox.
 
+    CommandLine:
+        xdoctest -m /home/joncrall/code/netharn/netharn/data/transforms/augmenters.py Resize:0
+
     Example:
         >>> from netharn.data.transforms.augmenters import *  # NOQA
         >>> img = demodata_hsv_image()
         >>> box = kwimage.Boxes([[.45, .05, .10, .05], [0., 0.0, .199, .199], [.24, .05, .01, .05]], format='xywh').to_tlbr()
-        >>> bboi = box.to_imgaug(shape=img.shape)
+        >>> bboi = box.to_imgaug(shape=tuple(img.shape))
         >>> self = Resize((40, 30))
         >>> aug1  = self.augment_image(img)
         >>> bboi1 = self.augment_bounding_boxes([bboi])[0]
+        >>> box1 = kwimage.Boxes.from_imgaug(bboi1)
+        >>> assert box1.br_y.max()  > 8
+        >>> assert tuple(bboi1.shape[0:2]) == (30, 40)
+        >>> assert tuple(aug1.shape[0:2]) == (30, 40)
 
     Example:
         >>> from netharn.data.transforms.augmenters import *  # NOQA
@@ -304,6 +311,18 @@ class Resize(augmenter_base.ParamatarizedAugmenter):
                                             self.target_size)
         return new_img
 
+    def _augment_bounding_boxes(self, bounding_boxes_on_images, random_state,
+                                parents, hooks):
+        # Fix for imgaug 0.4.0
+        return self._augment_bounding_boxes_as_keypoints(
+            bounding_boxes_on_images, random_state, parents, hooks)
+
+    def _augment_polygons(self, polygons_on_images, random_state, parents,
+                          hooks):
+        # Fix for imgaug 0.4.0
+        return self._augment_polygons_boxes_as_keypoints(
+            polygons_on_images, random_state, parents, hooks)
+
     def _augment_images(self, images, random_state, parents, hooks):
         self.target_size = None if self.target_size is None else np.array(self.target_size)
         return [self.forward(img, random_state) for img in images]
@@ -328,6 +347,7 @@ class Resize(augmenter_base.ParamatarizedAugmenter):
             >>> aug = self.augment_keypoints(keypoints_on_images)
             >>> assert np.all(aug[0].shape == self.target_size[::-1])
         """
+        print('keypoints_on_images = {!r}'.format(keypoints_on_images))
         result = []
         target_size = np.array(self.target_size)
         target_shape = target_size[::-1]