+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Fixed 
 * Handle "No running processes found" case in `XPU.coerce('auto')`
+* Resize now works with newer `imgaug` versions
 
 ### Changed 
 * Using new mode in `gpu_info`, this is more stable